Project

+++ Src/JavaSpringBoot/EcommerceBookSales/src/main/java/com/example/ecommercebooksales/entity/OrderItem.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer orderItemId;
+    private Long orderItemId;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
